added tests for utility module refactored various commands and services optimized code for message embeds removed emoteUtils completely and moved to service removed suggestion validation, because its now done via feature config fixed regex for discord links changed template to handle the case in which a member which suggested it originally left the server added suggestion update exception in cases the message to edit, does not qualify as a suggestion added check to not allow cross server quoting and added test case refactored mocking for jda objects

+++ abstracto-application/abstracto-modules/experience-tracking/experience-tracking-impl/src/test/java/dev/sheldan/abstracto/experience/ExperienceRelatedTest.java
@@ -23,17 +23,17 @@ public abstract class ExperienceRelatedTest {
         return experiences;
     }
 
-    protected AUserExperience getUserExperienceObject(AServer server, int i) {
-        AUserInAServer userObject = MockUtils.getUserObject((long)i, server);
+    protected AUserExperience getUserExperienceObject(AServer server, long i) {
+        AUserInAServer userObject = MockUtils.getUserObject(i, server);
         AExperienceLevel level = AExperienceLevel
                 .builder()
-                .level(i)
-                .experienceNeeded((long) i * 100)
+                .level((int)i)
+                .experienceNeeded(i * 100)
                 .build();
         return AUserExperience
                 .builder()
                 .user(userObject)
-                .experience((long) i)
+                .experience(i)
                 .currentLevel(level)
                 .build();
     }
